requirements.txt
torch
pydub
jieba
pypinyin
soundfile

# Openvino
openvino==2024.4.0
onnxruntime-openvino==1.20.0
